requirements.txt
# This file specifies the exact versions of dependencies for reproducibility.
# Use 'pip install -r requirements.txt' to install them.

pandas==2.2.3
numpy==2.1.3
vectorbt==0.28.0
pandas-ta==0.3.14b0
plotly==5.24.1
PyYAML==6.0.2
matplotlib==3.10.1
joblib==1.4.2